Split client and deployment sections
